+++ src/log.ts
@@ -1,5 +1,16 @@
-import * as types from './types';
 import { OutputChannel, window } from 'vscode';
+
+/**
+ * Log level enum (used for OutputChannel printing)
+ * 
+ * @export
+ * @enum {number} Log level
+ */
+export const enum LogLevel {
+    INFO,
+    WARNING,
+    ERROR
+}
 
 /** OutputChannel to print to in the main editor */ 
 let outputChannel: OutputChannel;
@@ -11,7 +22,7 @@ let outputChannel: OutputChannel;
  * @param {string} string Text to print in the OutputChannel
  * @param {types.LogLevel} [level] Log level to print (useful for output colorizr)
  */
-export function log(string: string, level?: types.LogLevel) {
+export function log(string: string, level?: LogLevel) {
     if (!outputChannel) {
         outputChannel = window.createOutputChannel('Mux');
     }